Moving styles for activities and changing EditProfile component like a view
