+++ src/assets/main.scss
@@ -9,3 +9,4 @@
 @import './styles/header';
 @import './styles/profile';
 @import './styles/vue-multiselect.min.css';
+@import './styles/activities';
