+++ src/assets/styles/_general.scss
@@ -67,12 +67,17 @@ body {
     border-radius: 10px;
   }
   #card-outline .card{
-      border-radius: 0;
+      border-radius: 15px;
       box-shadow: 5px 5px 5px #ffda00;
       transition: all 0.3s ease-in;
       -webkit-transition: all 0.3s ease-in;
       -moz-transition: all 0.3s ease-in;
     }
+  #card-outline .card:hover {
+    border-radius: 20px;
+    box-shadow: 5px 5px 10px #9e9e9e;
+    cursor: pointer;
+  }
   .line {
     padding: 1px;
     background-image: linear-gradient(to right, white, #F57C00, white);
