+++ src/views/Home.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="col home-content">
-    <b-alert show class="text-center" variant="success" dismissible fade>
+    <b-alert show class="text-center" variant="warning" dismissible fade>
       <b>Well done!</b> You successfully logged in to this website.
       <br>Welcome <b>{{ form.username }}</b>
     </b-alert>
